Define default metric e_requests_total as custom metric
