+++ runtime/appruntime/api/handler_test.go
@@ -21,13 +21,13 @@ import (
 	encore "encore.dev"
 	"encore.dev/appruntime/api"
 	"encore.dev/appruntime/config"
-	"encore.dev/appruntime/metrics"
 	"encore.dev/appruntime/metrics/metricstest"
 	"encore.dev/appruntime/model"
 	"encore.dev/appruntime/reqtrack"
 	"encore.dev/appruntime/trace"
 	"encore.dev/appruntime/trace/mock_trace"
 	"encore.dev/beta/errs"
+	usermetrics "encore.dev/metrics"
 )
 
 type mockReq struct {
@@ -362,11 +362,11 @@ func testServer(t *testing.T, klock clock.Clock, mockTraces bool) (*api.Server, 
 
 	logger := zerolog.New(os.Stdout)
 	testMetricsExporter := metricstest.NewTestMetricsExporter(logger)
-	metrics := metrics.NewManager(nil, cfg, logger)
 	rt := reqtrack.New(logger, nil, tf)
+	metricsRegistry := usermetrics.NewRegistry(rt, uint16(len(cfg.Static.BundledServices)))
 	json := jsoniter.ConfigCompatibleWithStandardLibrary
 	encoreMgr := encore.NewManager(cfg, rt)
-	server := api.NewServer(cfg, rt, nil, encoreMgr, logger, metrics, json, true, klock)
+	server := api.NewServer(cfg, rt, nil, encoreMgr, logger, metricsRegistry, json, true, klock)
 	return server, traceMock, testMetricsExporter
 }
